backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
httpx
beautifulsoup4
python-dotenv
celery
redis
python-multipart
python-jose[cryptography]
passlib[bcrypt]
